fix(seo): extract landing page into Server Component to enable page-level metadata
